requirements.txt
setuptools-rust==0.12.1
configparser
paramiko
python-hpilo
pprintpp
datetime
pyvmomi
